Validate operation variable types against usage
Summary:
We currently validate that variables are used at compatible locations *within* the span of a single document (e.g. query/fragment variable used within that query/fragment) and at fragment boundaries (checking args against argDefs). However, we don't currently check that usages of global (query) variables within fragments are valid - instead, we rely on query persistence to detect mismatches. For example, the following should fail but doesn't:

```
query Query($cond: Boolean) {
   ...Fragment
}
fragment Fragment on Query {
  me {
    name include(if: $cond) # <-- Should error, expects "Boolean!", got "Boolean"
  }
}
```

Because `$cond` is a global variable being used in fragment, we don't currently check it against the definition in its including queries.

This diff adds this validation, extending the responsibilities of our existing `skip_unused_variables` transform. That's now renamed to `validate_operation_variables`, and it handles both pruning unused variables as well as validating that the types of variable definitions match up with how variables are used. To find the types with which variables are used, the existing `InferVariablesVisitor` now tracks the most precise type with which each variable is used (before it just recorded the first type). Whereas before, `skip_unused_variables` just looked at the names of the used variables, we now also check their types (which we were recording anyway!).

One additional change is that `apply_fragment_arguments` wasn't preserving the true types of variables. It overwrote variable usages with the types from the variable definitions, which is now changed to keep an accurate picture of the truly required type of each variable usage.

Reviewed By: alunyov

Differential Revision: D28404016

fbshipit-source-id: 4a0f81be83e9543e79bfe36848ef0fbce2aa1f2d