move feature_flag defaulting to config loading
Summary:
Both, `ProjectConfig` and `Config` had a `feature_flags` property. The property from `Config` was just used in multiple places as a fallback to the `ProjectConfig.feature_flags`.

This diff removes the `Config.feature_flag` prop and moves the fallback behavior to the loading of the `ProjectConfig`.

Note that this does NOT alter config file behavior. The changed structs are loaded from a similar named serde-enabled struct with the `...File` suffix. Those are unchanged.

Reviewed By: alunyov

Differential Revision: D29817904

fbshipit-source-id: ecfb535990c9520c54ee0344f0fd957d4ed59982