Promote type refinement to default feature (no flag)
Summary:
Makes precise type refinement the default, without a feature flag. This has been the default behavior internally for some time and the behavior is stable. We've mentioned in release notes that this behavior will change in upcoming releases and encouraged users to try out the new behavior (both in release notes and on Slack).

Note that a few tests failed under the new mode, which uncovered a case that we did not yet handle when the feature is enabled.

RelayReader needs to know whether the fields of a fragment are expected to be present or not - ie, was there a type-mismatch in a parent or at the root that means child selections are not expected to exist. This is consequential since it affects `isMissingData`, which in turn determines whether we suspend on "missing" data or not. However, there's a catch with checking for type mismatches on the root record: we hard-code the root record type to `__Root` since we don't know at runtime what the schema calls it. Checking the type against the type in the AST will therefore generally fail, which would cause us to incorrectly decide data is expected to be missing. We currently handle this case by checking if the selector starts at ROOT_ID, but this does not work for mutations/subscriptions which generate unique root ids. This is acceptable in practice as we don't read data for mutations/subscriptions in a situation where we would use isMissingData to decide whether to suspend or not.

For now I'm leaving this behavior unchanged, since it's the behavior we actually observe at runtime when the feature flag is on. See the linked task for addressing this edge-case.

Reviewed By: jstejada

Differential Revision: D29917883

fbshipit-source-id: 55c9520ff472f6985cba5045c1612b8d78e46875