Remove MapByID subscriptions experiment
Summary:
This was an experimental alternative implementation for tracking which fragment subscriptions depended on which records in the store. The idea was to keep a mapping of RecordIDs to a set of subscribers interested in that ID. When notifying subscribers of changes, this would allow iterating just the changed IDs, finding the affected subscribers, and re-evaluating them. Our experiment results were mixed. While this did reduce time spent "notifying" subscribers, it also had other tradeoffs that we suspect are related to increased memory usage (the outer Map and all the inner Sets) as well as the time spent re-allocating / traversing those maps and the time spent removing entries.

Instead of this approach we've moved forward with other changes that minimize the number of times we have to notify subscribers.

Reviewed By: alunyov

Differential Revision: D29977497

fbshipit-source-id: 88bf87b3b781914d5b0576a196824ec78e57f9c5