Update the way with handle different paths in single project config file.
Summary:
In `SingleProjectConfig` we don't have an explicit `root` settings, so we were assuming that the `current_dir` is the `root`.

And the basic path normalization were just stripping the `root_dir` from configs paths (sources, schemas, artifacts). But this approach isn't working for cases where some paths are outside of `current_dir`. Like in the issue: https://github.com/facebook/relay/issues/3673

This diff fixes the issue by changing the way we detect `root_dir` for the project.

1) we create canonical paths from `current_dir` + `path`
2) we detect a common path between all config paths (src, schema, extensions, artifactsDirectory) - this common path is the `root_dir`.
3) we strip `root_dir` from all config paths when creating an instance of MultiProjectConfig.

Reviewed By: rbalicki2

Differential Revision: D32931262

fbshipit-source-id: f1940fc9ce975f7811c3f8f5a4d56790bd2bc74a