Stop @preloadable Typescript code generation from emitting an inline require (#3685)
Summary:
Fixes https://github.com/facebook/relay/issues/3683

I found that when I added the `preloadable` directive to a query in a TS project it would cause my application to crash at runtime because my bundler didn't transpire the `require` call that the `relay-compiler` generated in order to populate the `PreloadableQueryRegistry` map. This PR addresses this by changing it to an eager esmodule import when the code generation language is set to Typescript.

Pull Request resolved: https://github.com/facebook/relay/pull/3685

Test Plan:
* Tests still pass
* Locally built and utilized a version of relay-compiler in my minimal repro repo (https://github.com/vincentriemer/new-relay-compiler-issues-repro) and verified the `AppQuery.graphql.ts` generated output contained the eager esmodule import instead of an inline require.

Reviewed By: kassens

Differential Revision: D33064453

Pulled By: vincentriemer

fbshipit-source-id: 35fe3712f747a454297c17668a3203a9fd951410