Use WalkDir to find files if watchman is not available
Summary:
This change should fix https://github.com/facebook/relay/issues/3680.

The main cause of the issue is that canonical path generated in rust on windows starts with `\\?\`.

And the `glob` library we were using is explicitly checking for this prefix:

https://github.com/rust-lang-nursery/glob/blob/master/src/lib.rs#L179-L184

And later decides to return an empty result if the `pattern` starts with prefix:

https://github.com/rust-lang-nursery/glob/blob/master/src/lib.rs#L223-L235

We have two options:
- Try to remove window prefixes, etc, and make it work with `glob`. A hacky way, doesn't really work.

- Use `WalkDir` - we already were searching for all files in `root_dir` anyway. This is much cleaner.

Reviewed By: tyao1

Differential Revision: D33098204

fbshipit-source-id: e7d0c4538d55d98a076d65ee27ae8751d8d7380f