relay add readme files to packages (#3689)
Summary:
This diff is adding README.md to the package on NPM,

We have our compiler documentation in the compiler README file already, but it is not visible on NPM.

We could also update other packages' REAMDE files to highlight their purpose, and point to the correct place on the documentation website for a more detailed version.

But for now, let's just add basic files, and populate content later.

Pull Request resolved: https://github.com/facebook/relay/pull/3689

Test Plan:
Tested in the local github checkout, running `yarn gulp` and see how the README's are copied to the `dist`:

See:
{F686470070}

These REAME files should be rendered on NPM page, similar to our `relay-config` package:

https://www.npmjs.com/package/relay-config

**Static Docs Preview: relay**
|[Full Site](https://our.intern.facebook.com/intern/staticdocs/eph/D33072151/V5/relay/)|

|**Modified Pages**|

Reviewed By: voideanvalue

Differential Revision: D33072151

Pulled By: alunyov

fbshipit-source-id: 6629a0d023d8f167926a6e536b9de2ae0ee14c6c