Correctly extract module names on windows (#3694)
Summary:
So, my last attempt(https://github.com/facebook/relay/pull/3692) to fix the windows builds and incorrect name extraction didn't succeed. It turns out that when we were receiving a path from SourceLocation it contained `/` separators, and on windows, we were expecting this `\`.

Anyways, the new approach is to try to re-construct a `Path` from `&str`, check `file_stem` (if it's not `index` return it), otherwise return the last non-index part of the path.

Pull Request resolved: https://github.com/facebook/relay/pull/3694

Test Plan: - Tested on windows with an example project. Test project works.

Reviewed By: tyao1

Differential Revision: D33133403

Pulled By: alunyov

fbshipit-source-id: a05d8303ad4f4c783f5431545c31683937dfdb01