Move `ProjectConfig` and some other things to `relay-config`
Summary:
The goal of this change is to allow different crates to have access to the `project_config` structure. Currently, it's defined on the `relay-compiler` crate, so other `crates` that are dependencies of the relay-compiler cannot import the type of `ProjectConfig` from `relay-compiler`.

One of these crates that needs access to the `ProjectConfig` is `relay-transforms`. For now, we just passing pieces of the configuration to the `transforms` - things like project_name, feature flags, etc.

With this change, we will be able just to pass a reference to `&ProjectConfig`. An example of this is coming in the next diff.

Also, other pieces of `config` from the compiler can be moved to `relay-config`, and I'll finish this in follow-ups. For now, I want to keep the set of changes minimal.

Reviewed By: tyao1

Differential Revision: D33151243

fbshipit-source-id: 638529c37da33603ca5e20aae5a9b644908770e1