fix test utils for ResolverFragments
Summary:
Fix for https://github.com/facebook/relay/issues/3718

 ---

When we're bundling `relay-rutime` the bundle script puts all JS files in the `lib` directory.

So, when we're requiring this
`relay-runtime/store/ResolverFragments` module from `relay-runtime` it is not there. It's in the  `relay-runtime/lib/store/ResolverFragments`.

The proposed fix is to re-export `ResolverFragments` from `relay-runtime`.

Reviewed By: kassens

Differential Revision: D33454330

fbshipit-source-id: 2922e8eccfa51ff25873d3557a4febb56392d00d