Make compiler transforms customizable and applied as a Vec of Transforms (#3701)
Summary:
Some advanced usages of Relay benefit from customizing the compiler so that users can provide custom transformations to the documents to affect how artifacts are generated. This PR adds the capability to specify these transformations in the `ProjectConfig` as a vector of `Transform`s.

In order to support this flexibility, I refactored the `apply_transforms` module in `relay-transforms` to apply transforms from a vector, and create default vectors for each of the steps in the execution pipeline (common, reader, operation, typegen, etc). These are used when there's no configuration for custom transforms, and are also exposed in the `relay-transforms` crate so they can be used and customized before setting them in the configuration.

The `Transform` is a new struct that takes a name, and a transform function, which takes a `project_config`, the `program`, and `base_fragment_names`, and returns a `DiagnosticResult<Program>`. I created closures for each of the current transforms to accommodate to this signature (wrapping the result in an `Ok` whenever the transform returned just the `Program`). Maybe it could be useful to enforce this signature in the transforms themselves at some point, to avoid the closure.

Now we have a generic function that takes the transforms vector, and applies it in order. It also wraps each transform call to log the time with the perf logger, and conditionally prints stats. Given that only the normalization step was explicitly passing the configuration to print stats, I made it so that the other steps explicitly pass `None` to this param.

One thing to note, is that I made a slight modification in how stats are printed in the normalization step. Before, we were printing the stats before starting the process with the string "normalization start", but now, since we have a generic function that does this, this is printed with the step function name, so this will show up as "apply_normalization_transforms start". Hope this change is acceptable, otherwise we will need to pass an extra string param there.

Also, this is my first time using Rust, so there's probably a lot of things that could be done in a more efficient/idiomatic way.

_Note: This ended up being a big PR, I'm happy to split it up into smaller chunks if needed, just wanted to get feedback on the approach_

Pull Request resolved: https://github.com/facebook/relay/pull/3701

Reviewed By: kassens

Differential Revision: D33336957

Pulled By: alunyov

fbshipit-source-id: ec83c7c7aef204195769c466e3647d228b57ec40