Fix null comparison for skip_printing_nulls
In `skip_printing_nulls` (https://github.com/facebook/relay/pull/3626) we change some `null` values to be undefined by skipping them in the output.

Generally, the js code uses loose equality, either as
 * `field.X != null`
 * `field.X ?? default`
 * `if (field.X) {`

which allows us to do this safely.

The comparison this commit fixes was a rare exception to that showing up when using `@required`, mocking data and `skip_printing_nulls` is enabled.