Add a JSON file query persister strategy to relay-compiler (#3666)
Summary:
**NOTE**: This is essentially the first time I've developed in Rust so feel free to eviscerate it because...

![I have no idea what I'm doing](https://user-images.githubusercontent.com/1398555/144315265-3529ef64-6096-446f-91c0-07710ede5c34.jpg)

Previously in the new rust relay-compiler your only option for persisting queries is to use the remote persister which makes POST requests to a given endpoint with the operation text to be persisted. This PR adds the option to instead persist your queries to a local JSON file in a very similar way to how the JS compiler originally did it.

If you had a basic remote persister config like this:

```js
persistConfig: {
  url: "http://localhost:3000",
  params: {}
}
```

changing it to a local JSON file persister means updating the config to:

```js
persistConfig: {
  file: "./queries.json"
}
```

When this is enabled and a new query is persisted the compiler will first attempt to read the local file if it exists as a hash map (creating a new one if there is no existing data), hashes the operation text, adds the operation text to the hashmap keyed by the hashed version of the operation, and then serializes & writes the resulting map to the configured path.

Pull Request resolved: https://github.com/facebook/relay/pull/3666

Test Plan:
Leveraging my own [simple repo](https://github.com/vincentriemer/new-relay-compiler-issues-repro) I've been using for reproducing issues with the new relay compiler, installed the relay-compiler to a version including this PR's changes, and updated the `relay.config.js` file as such:

```diff
module.exports = {
  "src": "./src",
  "schema": "./schema.graphql",
  "artifactDirectory": "./src/__generated__",
  "language": "typescript",
  "persistConfig": {
-    "url": "http://localhost:3000",
-    "params": {}
+    filePath: "./queries.json"
  }
};
```

Then after running the relay compiler I verified that it created a `queries.json` file and it contained the expected persisted query json output:

```json
{
  "00dc69b2cfc4c1da3528e9aaf65a05091740fc71": "query AppQuery {\n  viewer {\n    submission(name: \"t3_qvyl2j\") {\n      ...Post_submission\n      id\n    }\n  }\n}\n\nfragment Post_submission on Submission {\n  title\n  author\n  selftext {\n    html\n  }\n}\n"
}
```

Reviewed By: josephsavona

Differential Revision: D32802562

Pulled By: alunyov

fbshipit-source-id: 8e577e37534daded65567274897e7218b92418fe