Fix properties being duplicated in typegen.
This happened when there was a combination between a concrete type with
a selection that overlapped with selections on an interface, that is
also implemented by the concrete type. In such cases, the base selections
are added to the selections of the concrete type, but that could result
in duplicates. Now the selections are merged instead.