Refactor RelayModernEnvironment to access the network via public methods
Summary:
This is small refactor of `RelayModernEnvironment` so it uses its public `getNetwork` method to access the configured network, so any subclasses can override it.

This is necessary because the network property is private and it's inaccessible to subclasses. Alternatively this could be achieved using `__network` instead, but I thought this was cleaner.

Reviewed By: kassens

Differential Revision: D34621426

fbshipit-source-id: 3184896b064a1aaed0e4185f1e88f4ae0db17dc7