Add test case for @module and provided variables (1/2)
Summary:
Found a bug in the compiler output when `module` fragments use provided variables.

When writing `apply_fragment_arguments` transform, I made the incorrect assumption that the only source of `SplitOperation`s are `no-inline` fragments. However, `SplitOperations` are generated by other directives as well.

 ---
This diff adds test cases with incorrect output (that will be fixed in the next diff).
Since `module` uses `no-inline` internally when `no-inline` is enabled for a fragment, the FeatureFlag for `no-inline` needed to be set to limited to produce the incorrect output.

Happy to introduce a new enum (something like `EnabledExcept`), but we shouldn't change our config structure for a test case.

Reviewed By: josephsavona

Differential Revision: D34645169

fbshipit-source-id: e22fef80d6fe817c1925bad9a92c06644f3ff352