Modify how RelayModernEnvironment accesses its network property to allow subclassing (re-land)
Summary:
NOTE: This is a re-application of D34621426 (https://github.com/facebook/relay/commit/32d17fa8e8928823d2ad9d28b53af7950d993a4a), which was reverted as part of a stack but was good.

This is small refactor of `RelayModernEnvironment` so it uses its public `getNetwork` method to access the configured network, so any subclasses can override it.

This is necessary because the network property is private and it's inaccessible to subclasses. Alternatively this could be achieved using `__network` instead, but I thought this was cleaner.

Reviewed By: dmitryrykun

Differential Revision: D34681535

fbshipit-source-id: 8959abd22052dc7b8b713e1bd40e74f4904c055a