Cache promises in fetchQueryInternal
Summary:
Cause fetchQueryInternal, which already maintains a cache for active requests, to cache the Promise to be thrown for that request. This allows us to re-throw the same promise each time without having an additional level of caching. The cache is cleaned up when the request is finished or unsubscribed.

This fixes a bug in the new React Cache implementation of useFragment that I discovered when writing additional tests for it.

Reviewed By: josephsavona

Differential Revision: D34569982

fbshipit-source-id: 676163bc577ab667de83044b47a0d5443c34a4d3