Combinatorial testing of useLazyLoadQuery
Summary:
Trying out new approaches to testing. Here we have a single test function with a configurable setup and configurable assertions. We manually list out each combination of inputs and then what should happen.

Even though this is not clever at all, it's testing most of the behavior that QueryResource-test was testing (haven't gotten to all of it yet), plus more, in 1/5 the lines of code. And instead of nearly-same code repeated over and over unmaintainably, it's mostly just a table of expected inputs and outputs. The actual code part is very short.

The only slightly-clever thing it does is that it tests every possible outcome for the fetch response — synchronous or async, error or success — but *only if* a fetch is expected to happen at all.

I'm still exploring other options for smarter testing (this approach isn't going to scale much beyond what I have here), but this approach at least unblocks the React Cache work.

Reviewed By: josephsavona

Differential Revision: D34494586

fbshipit-source-id: f0a0dbb6aae6b87af97fd9d02bdbe4e073169963