Update merging pending sources to fix slow incremental build
Summary:
## Context
The incremental build currently is not working 100% times. When merging pending sources, we always append a `doc_block`(empty array when there is none in the source) to a pending source. Then our `has_breaking_schema_change` check will always fail, because there is pending `doc_block`  and  `is_incremental` is set to false, we do a full build.

Why do we need to append the empty `doc_block`?
We need to have a way to tell that a source is empty when we add new pending sources. This is because a source can be deleted, and the processed sources still contains a stale value for the source, and we need to have the compiler remove that in incremental build.

## This diff
Adds some extra processing when we add a source to pending sources: When the pending source is empty and processed sources don't contain the entry, there is nothing to delete. So we don't have to add that to pending sources. Then the `has_breaking_schema_change` won't break.

Reviewed By: rbalicki2

Differential Revision: D35433967

fbshipit-source-id: dffdf1a89035c955cfbce709d605b1a264b882bf