Validate `rootFragment` to make sure it exists
Summary:
This diff changes the` build_schema` method in `relay-compiler`.

Now, `build_schema` accepts the slice of the `ExecutableDocuments` (not IRs) which can be used in the Resolver Parser to validate used fragment names (to make sure they are defined).

Later, we can use these `ExecutableDocuments` to extract `argDefs` from the fragments to use these are resolver field arguments.

This diff doesn't pass these `ExecutableDocuments` to `relay-lsp` (will be added a one the follow-ups)

A few questions:
- Is there a requirement that the fragment `onType` on the root fragment?  Should they have the same type? Or it's fine for them to be different types?
- What's `edgeTo`?
- Can we use `rootFramgnet` fragments that are defined outside of the relay resolver module?

Reviewed By: captbaritone

Differential Revision: D35406758

fbshipit-source-id: 9aa42af350253a75f8ca013ccf677b5beb7e5c29