Initial validation for the resolver fields
Summary:
This version is disabling the usage of fragment spreads in Relay Resolver Fragments. We currently do not have a Runtime API to correctly read/track updates in these fragments, so it easier to just throw an error here, and see if people have the use-case for it, and we then can prioritize and add the support for it.

Some Implementation Details:
 - We iterate over all fields in the schema and check field directives. This allows us to validate all resolver fields, not only those that are used in the code.
- We extract fragment names from these fields, and running the validation only for the Resolver's fragments.

What's next:
 - Add a validation that disable the usage of global variables in fragments.

Reviewed By: captbaritone

Differential Revision: D35679422

fbshipit-source-id: 2d808b810020201afa7a511bf95485ad4794e6a3