Add Location to InlineFragment IR.
Summary:
Currently InlineFragment IR doesn't contain location (span), and when there are errors related to the fragments, it may be hard to find where it's location.

This diff only adds the location to the fragment, it doesn't fix error messages.

Many InlineFragment IRs are generated, thus the location is set to generated.

Currently the only source of location is from `graphql-ir/src/build.rs`.

Reviewed By: rbalicki2

Differential Revision: D35758791

fbshipit-source-id: 64c2890d5ed39e111da60d99719ffa9ec9bcac3f