artifact_content refactoring 1/N: empty line after indirect 3d metadata annotation
Summary:
When I added the indirect 3d metadata, I missed that `generate_operation` conditionally added an empty line to separate the metadata annotations. This diff "fixes" that.

The rest of the diff stack does a bunch of refactoring to make the code easier to read and update in the future.

Reviewed By: rbalicki2

Differential Revision: D36097919

fbshipit-source-id: 7034431090df06c15592b5c95b9416a27e0342e4