artifact_content refactoring 2/N: introduce content sections
Summary:
Introducing content sections to make it easier to identify sections in artifact files and make code easier to read and write.

Intentionally added a bunch of temporary hacks in order to make this a no-op diff for artifact outputs. Will remove them later in the stack after some more refactoring in subsequent diffs.

Reviewed By: kassens

Differential Revision: D36097920

fbshipit-source-id: ff1e986d8bff0b4acb9d8ef7758b0caebea128eb