artifact_content refactoring 1.5/N: add comment markers for sections
Summary:
It was getting really hard to figure out where inconsistent newlines were coming from. Sprinkling a bunch of comments for where sections begin and end to make them easier to spot.

I quite like having the comments now that I've looked at this file so many times. Happy to remove the comments at the end of the stack if anyone asks.

Reviewed By: kassens

Differential Revision: D36125196

fbshipit-source-id: a80d598c644bee4c94ea1b1e3a87a56941afec61