Legacy behavior option
Summary:
Adds an option, controlled by another Feature Flag, to emulate the old behavior of retaining until a timeout on suspense and for the lifecycle of the owner component rather than its ancestor Cache. This is to allow existing apps to not have a behavior change right away.

Maybe this should be controlled by an Environment or Query-level setting rather than a GK?

Reviewed By: josephsavona

Differential Revision: D35748157

fbshipit-source-id: bb85dde0987c29dc1d7f5123e660e632d5de8bb4