Refactor existing directives to use @static
Summary:
This is the last set of changes related to `static`. There are some notable directives highlighted from the task that `static` would be helpful in:

* `@/match`
* `@/preloadable`
* `@/required`

Also removed tests from `@/required` and `@/match` that duplicate the efforts of `static`.

And to help with refactoring, some new unpacking/unwrapping mechanisms were introduced. See below for more details!

Reviewed By: rbalicki2

Differential Revision: D35916559

fbshipit-source-id: c024246f88a9f9eeb212ec92fd6018cfadc748db