Avoid bug where watch mode can get stuck in invalid state (#3922)
Summary:
1. Start the Relay compiler in watch mode
2. Find a file with an existing fragment and duplicate is (cmd+c, cmd+v in VSCode file tree)
3. Observe a compiler error in your terminal due to duplicate fragment names
4. Delete the `... copy.js` file

Compiler compiles with no errors

Compiler still complains about duplicate fragments despite the `... copy.js` file not existing any more

When merging new sets of pending files, we ignore empty (deleted in this case) files that are missing from the processed set of files. But we fail to consider the fact that the file might already exist in the pending set.

If a file is already in the pending set, always update it. Otherwise we can end up with state data in the compiler.

Pull Request resolved: https://github.com/facebook/relay/pull/3922

Reviewed By: voideanvalue

Differential Revision: D36629649

Pulled By: captbaritone

fbshipit-source-id: a92eb6cc92696a5af782950beed2dbbe340cda13