Report schema location for invalid types (#3924)
Summary:
With Client Edges, if you specify an `edge_to` with an invalid type name we were previously reporting an error at a generated location.

<img width="1124" alt="Screen Shot 2022-05-25 at 8 39 24 AM" src="https://user-images.githubusercontent.com/162735/170302170-2bf3004d-e118-4c2b-a863-fd2d899a6555.png">

This also ensures we provide a correct error location when you type a field type in a client schema extension.

Pull Request resolved: https://github.com/facebook/relay/pull/3924

Test Plan:
Create a Relay Resolver with `edge_to Foo` and see an error message that points to the correct location:

<img width="1126" alt="Screen Shot 2022-05-25 at 8 28 44 AM" src="https://user-images.githubusercontent.com/162735/170302024-342524e0-8a64-4afb-a2a9-ea676f2eee11.png">

Reviewed By: voideanvalue

Differential Revision: D36667389

Pulled By: captbaritone

fbshipit-source-id: 700414188c75f3043c889cf3fa7dce6841168da7