Remove {disable,enable}StoreUpdates from new useFragment return
Summary:
The legacy hook useFragmentNode() returns an object with the result data plus functions to disable and enable listening for store updates. These extra functions are *only* used in `useBlockingPaginationFragment()`, which is itself rarely used.

This means that we're allocating extra objects and closures for every single query/fragment hook invocation, only to discard most of that extra information. Let's optimize for the majority case and find an alternative for useBlockingPaginationFragment.

Reviewed By: davidmccabe

Differential Revision: D36647742

fbshipit-source-id: f28363c8608f00d0ef21280d872bb295a3030749