Update Compiler Playground tests (#3927)
Summary:
The compiler playground does not currently build due to the version of parkling_lot that we are using being incompatible with wasm-pack. I manually updated the parking_lot version to `0.12` in order to get it building again. This uncovered some tests which are currently failing. Updating these tests will make it easier to get the the compiler playground building again once we upgrade parking_lot officially.

Currently upgrading is somewhat complicated because we try to only have one version of each dependency installed at any given time in the Meta monorepo.

Pull Request resolved: https://github.com/facebook/relay/pull/3927

Reviewed By: josephsavona

Differential Revision: D36789186

Pulled By: captbaritone

fbshipit-source-id: 448502cdee3dd209c6f053e9cb236420e92dfc32