Support CommonJS for @no_inline (#3923)
Summary:
The goal of this PR is to support `no_inline` in CommonJS, but I did a little clean up along the way.

We used to clone the `no_inline` directive to all of the FragmentSpreads that used that source fragment. That felt like a hack so I switched it over to use `NoInlineMetadata` which uses the cool and very hip associated data stuff.

Pull Request resolved: https://github.com/facebook/relay/pull/3923

Reviewed By: josephsavona

Differential Revision: D36820179

Pulled By: captbaritone

fbshipit-source-id: e69449410f4e06b0b213a59894cee96342eca049