Persist explorer state in the URL/local storage (#3930)
Summary:
This should allow the compiler explorer to be used as a mechanism for sharing minimal reproducible compiler bugs and other examples of compiler behavior.

The general idea is to consolidate all the state in the compiler explorer into a single object (Redux style):

* Schema
* Document
* Feature Flags
* Language
* Selected tab

And then (json) serialize and compress that state and persist it to the URL hash and local storage. Note that by using the URL hash we ensure the content of the explorer is never sent to the server, and thus the privacy of the content is improved.

Pull Request resolved: https://github.com/facebook/relay/pull/3930

Test Plan:
https://user-images.githubusercontent.com/162735/171328459-a675c8e5-a9a6-4a27-af2e-1230245b84d2.mov

**Static Docs Preview: relay**
|[Full Site](https://our.intern.facebook.com/intern/staticdocs/eph/1654205352/relay/)|

|**Modified Pages**|

**Static Docs Preview: relay**
|[Full Site](https://our.intern.facebook.com/intern/staticdocs/eph/D36845080/V3/relay/)|

|**Modified Pages**|

Reviewed By: alunyov

Differential Revision: D36845080

Pulled By: captbaritone

fbshipit-source-id: 532cdbed4e0b1014ff3232e1b8aa47dce16cafa1