Avoid useMemo for fragment data except on plurals
Summary: Micro-optimization to avoid an unnecessary `useMemo` call in the common case. Previously the new useFragmentInternal_REACT_CACHE() hook memoized the translation from internal hook to fragment result. However, for singular fragments this is a simple property lookup and isn't worth memoizing. Here, we exploit the fact that plurality is a static property of the fragment (doesn't change dynamically for a particular fragment invocation) and only memoize the computation for plural fragments.

Reviewed By: davidmccabe

Differential Revision: D36796332

fbshipit-source-id: 5651e62f9c2502f1a634cc825719e3fe4ed7fefd